Add files via upload
The goal of this project is to analyze data to better understand what type of apps are likely to attract more users. In this project, we are hypothetically working as data analysts for a company that builds Android and iOS mobile apps, and our job is to enable developers to make data-driven decisions with respect to the kind of apps they build.